+++ app/layout.jsx
@@ -1,3 +1,5 @@
+import '../styles/globals.css';
+
 export const metadata = {
 	title: 'Home',
 	description: 'Welcome to Next.js',
