+++ utils/cache/itemIds.js
@@ -1,7 +1,7 @@
 import cache from 'memory-cache';
 import getAllItem from '../db/getAllItem';
 
-module.exports = async function RelevantItems() {
+export default async function RelevantItems() {
 	const allItems = await getAllItem();
 	return cache.put('allItems', allItems);
-};
+}
