+++ utils/cache/relevantItems.js
@@ -1,7 +1,7 @@
 import cache from 'memory-cache';
 import getAllRelevantItemInfo from '../db/getAllRelevantItemInfo';
 
-module.exports = async function RelevantItems() {
+export default async function RelevantItems() {
 	const relevantItems = getAllRelevantItemInfo();
 	return cache.put('relevantItems', relevantItems, 8.64e7 * 3);
-};
+}
