+++ utils/cache/updateItemIds.js
@@ -1,5 +1,5 @@
 import cache from 'memory-cache';
 
-module.exports = async function UpdateItemIds(updatedItemSet) {
+export default async function UpdateItemIds(updatedItemSet) {
 	return cache.put('allItems', updatedItemSet);
-};
+}
