+++ utils/db/getRelevantConsumableItemInfo.js
@@ -1,10 +1,10 @@
 import Item from '../../models/item';
 
-module.exports = async function getAllConsumableItemInfo() {
+export default async function getAllConsumableItemInfo() {
 	try {
 		const results = await Item.find({ itemClass: 'Consumable' });
 		return results;
 	} catch (error) {
 		return error;
 	}
-};
+}
