+++ utils/db/getRelevantGemItemInfo.js
@@ -1,10 +1,10 @@
 import Item from '../../models/item';
 
-module.exports = async function getAllGemItemInfo() {
+export default async function getAllGemItemInfo() {
 	try {
 		const results = await Item.find({ itemClass: 'Gem' });
 		return results;
 	} catch (error) {
 		return error;
 	}
-};
+}
