+++ utils/db/getRelevantGlyphItemInfo.js
@@ -1,10 +1,10 @@
 import Item from '../../models/item';
 
-module.exports = async function getAllGlyphItemInfo() {
+export default async function getAllGlyphItemInfo() {
 	try {
 		const results = await Item.find({ itemClass: 'Glyph' });
 		return results;
 	} catch (error) {
 		return error;
 	}
-};
+}
