+++ utils/formatData/cache/auctionHouse.js
@@ -1,10 +1,10 @@
-module.exports = async function AuctionHouse(data) {
-  let auctionHouses = {};
-  data.auctions.forEach(auctionHouse => {
-    auctionHouses[auctionHouse.name.toLowerCase().replace(/\s+/g, '-')] = {
-      id: auctionHouse.id,
-      name: auctionHouse.name,
-    }
-  })
-  return auctionHouses;
+export default async function AuctionHouse(data) {
+	let auctionHouses = {};
+	data.auctions.forEach((auctionHouse) => {
+		auctionHouses[auctionHouse.name.toLowerCase().replace(/\s+/g, '-')] = {
+			id: auctionHouse.id,
+			name: auctionHouse.name,
+		};
+	});
+	return auctionHouses;
 }
