+++ utils/formatData/props/auctionHouse.js
@@ -1,7 +1,7 @@
-module.exports = async function AuctionHouse(data) {
-  let auctionHouses = {};
-  Object.keys(data).forEach(auctionHouse => {
-    auctionHouses[auctionHouse] = data[auctionHouse].name;
-  })
-  return auctionHouses;
+export default async function AuctionHouse(data) {
+	let auctionHouses = {};
+	Object.keys(data).forEach((auctionHouse) => {
+		auctionHouses[auctionHouse] = data[auctionHouse].name;
+	});
+	return auctionHouses;
 }
