+++ utils/formatData/props/realm.js
@@ -1,7 +1,7 @@
-module.exports = async function Realm(data) {
-  let realmData = {};
-  Object.keys(data).forEach(realm => {
-    realmData[realm] = data[realm].name;
-  });
-  return realmData;
+export default async function Realm(data) {
+	let realmData = {};
+	Object.keys(data).forEach((realm) => {
+		realmData[realm] = data[realm].name;
+	});
+	return realmData;
 }
